LayoutTests/fast/mediastream/MediaStream-page-muted-expected.txt

Test enabling/disabling mock media capture devices

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



*** Mock capture devices should be enabled by default
PASS mediaStream is an instance of Object
PASS mediaStream.getTracks().length is 2

*** Muting capture devices
EVENT: mute
PASS event.target.muted is true

EVENT: mute
PASS event.target.muted is true


*** Unmuting capture devices
EVENT: unmute
PASS event.target.muted is false

EVENT: unmute
PASS event.target.muted is false

PASS successfullyParsed is true

TEST COMPLETE
